'fix version issue'

+++ my_sd/model_loader.py
@@ -12,17 +12,17 @@ import model_converter
 def preload_models_from_standard_weights(ckpt_path, device):
     state_dict = model_converter.load_from_standard_weights(ckpt_path, device)
 
-    encoder = VAE_Encoder().to(device=device, dtype=torch.float16)
+    encoder = VAE_Encoder().to(device=device)
     encoder.load_state_dict(state_dict['encoder'], strict=True)
 
 
-    decoder = VAE_Decoder().to(device=device, dtype=torch.float16)
+    decoder = VAE_Decoder().to(device=device)
     decoder.load_state_dict(state_dict['decoder'], strict=True)
 
-    diffusion = Diffusion().to(device=device, dtype=torch.float16)
+    diffusion = Diffusion().to(device=device)
     diffusion.load_state_dict(state_dict['diffusion'], strict=True)
 
-    clip = CLIP().to(device=device, dtype=torch.float16)
+    clip = CLIP().to(device=device)
     clip.load_state_dict(state_dict['clip'], strict=True)
 
     return {
